🐛 move CoilTypeEnumMeta and CoilType to a new file
