+++ bluemira/equilibria/coils/_grouping.py
@@ -23,13 +23,14 @@ if TYPE_CHECKING:
 import numpy as np
 
 from bluemira.base.look_and_feel import bluemira_warn
-from bluemira.equilibria.coils._coil import Coil, CoilType
+from bluemira.equilibria.coils._coil import Coil
 from bluemira.equilibria.coils._field import CoilGroupFieldsMixin, CoilSetFieldsMixin
 from bluemira.equilibria.coils._tools import (
     _get_symmetric_coils,
     check_coilset_symmetric,
     get_max_current,
 )
+from bluemira.equilibria.coils.coil_type import CoilType
 from bluemira.equilibria.constants import I_MIN
 from bluemira.equilibria.error import EquilibriaError
 from bluemira.equilibria.grid import Grid
